requirements/base.txt
psycopg2==2.7.3.2
bcrypt==3.1.4

django-compressor==2.2
django-model-utils==3.0.0
django-forms-bootstrap==3.1.0
Django==2.0.1
pytz
raven==6.4.0
django-jsonify==0.3.0

django-settings-context-processor==0.2
django-admin-decorators==0.1
django-webpack-loader==0.5.0

djangorestframework==3.7.7

django-redis==4.8.0
redis==2.10.6
celery==4.1.0

djangorestframework-jwt==1.11.0
django-rest-auth[with_social]==0.9.2

ua-parser==0.7.3
user-agents==1.1.0
django-user-agents==0.3.2

elasticsearch==6.1.1
django-rest-elasticsearch==0.3.4
elasticsearch-dsl==6.1.0

Pillow==5.0.0
